+++ src/Libraries/SmartStore.Services/Search/LinqCatalogSearchService.cs
@@ -417,10 +417,13 @@ namespace SmartStore.Services.Search
 				}
 				else if (filter.FieldName == "available")
 				{
-					query = query.Where(x => !(
-						x.StockQuantity <= 0 && x.BackorderModeId == (int)BackorderMode.NoBackorders &&
-						(x.ManageInventoryMethodId == (int)ManageInventoryMethod.ManageStock || x.ManageInventoryMethodId == (int)ManageInventoryMethod.ManageStockByAttributes)
-					));
+					// We cannot include ManageInventoryMethod.ManageStockByAttributes because it's only functional with MergeWithCombination.
+					var manageStock = (int)ManageInventoryMethod.ManageStock;
+
+					query = query.Where(x => 
+						x.ManageInventoryMethodId != manageStock ||
+						(x.ManageInventoryMethodId == manageStock && (x.StockQuantity > 0 || x.BackorderModeId != (int)BackorderMode.NoBackorders))
+					);
 				}
 				else if (filter.FieldName.StartsWith("price"))
 				{